teacher_directory/teacher_directory/urls.py
"""teacher_directory URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/3.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.conf.urls.static import static
from django.conf import settings
from django.contrib import admin
from django.urls import path, re_path, include
from rest_framework_nested import routers
from rest_framework.authtoken import views

from teachers.views import TeachersView, TeachersViewSet

router = routers.DefaultRouter()
router.register('teachers-rest', TeachersViewSet, basename="teachers-rest")
urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
    path('api-token-auth/', views.obtain_auth_token, name='api-token-auth'),
    path('teachers/', TeachersView.as_view(), name="teachers-list"),
    path('teachers/<int:teacher_id>/', TeachersView.as_view(), name="test"),
    re_path(r'^', include(router.urls)),
]
urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
