teacher_directory/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-rest-framework-nested-resource==1.3.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3